remove EntryManager to unify with LsCommand
